Added texts for the UI and comments for code
Added a text for when the program is downloading, in my usage I found that when downloading something huge it just appears to be frozen when it's still downloading. The text would help potential users know it's attempting to download.

Added a try catch for invalid inputs for either the file path or link box, also comes with a helpful reminder text in the UI.

Added some comments to the code to help understand what does what.